Split encoding into multiple files, rename id to voteId
