+++ src/logic/proposals/encoding/index.ts
@@ -1,0 +1,4 @@
+export * from './encoding';
+export * from './metadata';
+export * from './proposal-id';
+export * from './types';
